test(generator): change test
