+++ src/test/java/lotto/domain/generator/LottoGeneratorTest.java
@@ -1,5 +1,6 @@
 package lotto.domain.generator;
 
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
@@ -20,6 +21,22 @@ class LottoGeneratorTest {
         assertAll(
                 () -> assertThat(generatedLottos.size()).isEqualTo(8),
                 () -> assertThat(generatedLottos.getLotto(0).size()).isEqualTo(6)
+        );
+    }
+
+    @Test
+    @DisplayName("넣은 돈이 형식에 맞는지 확인하는 테스트")
+    void REQUEST_TYPE_VALIDATION_CHECK() {
+        MoneyRequest moneyRequest1 = new MoneyRequest(1);
+        MoneyRequest moneyRequest2 = new MoneyRequest(1200);
+
+        assertAll(
+                () -> assertThatThrownBy(() -> new LottoGenerator(moneyRequest1))
+                        .isInstanceOf(IllegalArgumentException.class)
+                        .hasMessageContaining("최소 구매금액은 1000원 입니다."),
+                () -> assertThatThrownBy(() -> new LottoGenerator(moneyRequest2))
+                        .isInstanceOf(IllegalArgumentException.class)
+                        .hasMessageContaining("1000원 단위로 입력해주세요.")
         );
     }
 }
